docs(pr-review): Add comprehensive PR #14 fix plan and restructure analytics
- Add detailed PR_14_FIXES.md documenting 42 issues from CodeRabbit review
- Restructure analytics pages from /analytics to /dashboard routes
- Remove unused dashboard components and charts
- Update navigation and routing structure
- Document critical fixes needed for database safety, accessibility, and i18n

Note: TypeScript errors are expected and documented in PR_14_FIXES.md for next PR
