+++ src/app/[locale]/(app)/dashboard/products/page.tsx
