+++ src/components/charts/index.ts
@@ -1,8 +1,6 @@
 // Optimized chart exports to improve tree shaking
 // Import only the specific chart components needed
 
-export { VisitorsChart } from '../dashboard/VisitorsChart'
-
 // Future chart components can be added here with lazy loading
 // export const PieChart = dynamic(() => import('./PieChart'), { ssr: false })
 // export const BarChart = dynamic(() => import('./BarChart'), { ssr: false })
